requirements.txt
attrs==22.2.0
aws-cdk.assets==1.199.0
aws-cdk.aws-applicationautoscaling==1.199.0
aws-cdk.aws-autoscaling-common==1.199.0
aws-cdk.aws-cloudwatch==1.199.0
aws-cdk.aws-codeguruprofiler==1.199.0
aws-cdk.aws-codestarnotifications==1.199.0
aws-cdk.aws-ec2==1.199.0
aws-cdk.aws-ecr==1.199.0
aws-cdk.aws-ecr-assets==1.199.0
aws-cdk.aws-efs==1.199.0
aws-cdk.aws-events==1.199.0
aws-cdk.aws-iam==1.199.0
aws-cdk.aws-kms==1.199.0
aws-cdk.aws-lambda==1.199.0
aws-cdk.aws-logs==1.199.0
aws-cdk.aws-s3==1.199.0
aws-cdk.aws-s3-assets==1.199.0
aws-cdk.aws-signer==1.199.0
aws-cdk.aws-sns==1.199.0
aws-cdk.aws-sqs==1.199.0
aws-cdk.aws-ssm==1.199.0
aws-cdk.cloud-assembly-schema==1.199.0
aws-cdk.core==1.199.0
aws-cdk.cx-api==1.199.0
aws-cdk.region-info==1.199.0
cattrs==22.2.0
constructs==3.4.293
importlib-resources==5.12.0
jsii==1.80.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.5.0
